Checkpoint for searching multiple genes: UI in place, but only the first is used
